refactor(app): move actions-page template to be within nested app/ folder
This will encourage to understand that there are some templates intended to be rendered within app
and others that not
