+++ src/templates/app/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as ActionsPage } from './actions-page'
